# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/calendar_employees:
    get:
      tags:
        - CalendarEmployees
      summary: List CalendarEmployees
      description: Get a paginated list of CalendarEmployees with optional filtering and sorting
      operationId: listCalendarEmployees
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEmployeeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CalendarEmployees
      summary: Create CalendarEmployee
      description: Create a new CalendarEmployee
      operationId: createCalendarEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarEmployeeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEmployee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/calendar_employee_statuses:
    get:
      tags:
        - CalendarEmployeeStatuses
      summary: List CalendarEmployeeStatuses
      description: Get a paginated list of CalendarEmployeeStatuses with optional filtering and sorting
      operationId: listCalendarEmployeeStatuses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEmployeeStatusList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CalendarEmployeeStatuses
      summary: Create CalendarEmployeeStatus
      description: Create a new CalendarEmployeeStatus
      operationId: createCalendarEmployeeStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarEmployeeStatusRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEmployeeStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    CalendarEmployee:
      type: object
      required:
        - id
        - calendar_id
        - employee_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        calendar_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateCalendarEmployeeRequest:
      type: object
      required:
        - calendar_id
        - employee_id
        - metadata
      properties:
        calendar_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CalendarEmployeeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CalendarEmployee'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CalendarEmployeeStatus:
      type: object
      required:
        - id
        - calendar_id
        - employment_status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        calendar_id:
          type: string
          format: uuid
        employment_status:
          type: string
        metadata:
          type: object
...
    CreateCalendarEmployeeStatusRequest:
      type: object
      required:
        - calendar_id
        - employment_status
        - metadata
      properties:
        calendar_id:
          type: string
          format: uuid
        employment_status:
          type: string
        metadata:
          type: object
...
    CalendarEmployeeStatusList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CalendarEmployeeStatus'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []